Improving performance of complex number class by allowing mutability.
Times for old implementation, previous revision, and current:
ClientVM: 2s, 9s, 5s
ServerVM: 1s, 5s, 2s

So this commit does not get back all the performance that was lost in the last commit.  I'm not sure if it's possible to get back the remaining performance without manually inlining these methods.

git-svn-id: file:///home/pauldoo/Programs/svnrepo@307 8e395407-42ff-0310-94d9-ec183df6106f